Account for move of meta_data to dmod.core.
