+++ python/lib/modeldata/dmod/test/test_data_requirement.py
@@ -1,5 +1,5 @@
 import unittest
-from ..modeldata.data.meta_data import DataDomain, DataFormat, DataRequirement, DataCategory, DiscreteRestriction
+from dmod.core.meta_data import DataDomain, DataFormat, DataRequirement, DataCategory, DiscreteRestriction
 
 
 class TestDataRequirement(unittest.TestCase):
