+++ python/lib/modeldata/setup.py
@@ -14,7 +14,7 @@ setup(
     author_email='',
     url='',
     license='',
-    install_requires=['numpy>=1.20.1', 'pandas', 'geopandas', 'dmod-communication>=0.4.2', 'minio',
+    install_requires=['numpy>=1.20.1', 'pandas', 'geopandas', 'dmod-communication>=0.4.2', 'dmod-core>=0.1.0', 'minio',
                       'hypy@git+https://github.com/noaa-owp/hypy@master#egg=hypy&subdirectory=python',
                       ],
     packages=find_namespace_packages(exclude=('tests', 'schemas', 'ssl', 'src'))
